Audrius Butkevicius' "Counters" patchset.
Yeah I couldn't figure out how to cleanly apply the patches in
the format they were sent to the libssh list here:

  http://www.libssh.org/archive/libssh/2014-01/0000023.html

but with some manual tinkering was able to end up with this.
The last patch in particular wouldn't apply cleanly, possibly
because it's based on master (this branch is based off v0-6).