kex: fix first_kex_packet_follows wrong-guess case
Fix treatment of 'first_kex_packet_follows':

In a previous change, an issue was fixed for the case that an initial
KEXINIT message had its 'first_kex_packet_follows' field set.  That
field must be used by both client and server when computing a given
session's ID, or the disagreement will lead to errors like "Bad
hostkey signature" when using modern dropbear clients.

That fix was correct for the case that the KEXINIT message did indeed
include a correct guess as to the preferred key exchange algorithm
('curve25519-sha256@libssh.org' by default in both libssh on the
server-side and dropbear on the client-side).

But, it did not include logic to ignore the first KEX_DHINIT message
for the case that the guess was actually wrong.  With this change,
such messages are now properly ignored, and I observe dropbear
clients are working when 'curve25519-sha256@libssh.org' is disabled
in libssh on the server-side.