Add support for SHA2 algorithms
This patch adds support for both hmac-sha2-256 and hmac-sha2-512 HMAC
algorithms in libssh. This makes the HMAC algorithm configurable in the
ssh_crypto_struct so it can be set to the appropriate algorithm.

This also fixes the stretching algorithm for the generated shared hmac
secret to work for all types and not just a special for the in_cipher
and out_cipher keys.

In a few other places there were hardcoded assumptions about sha1 always
being 20 bytes in length, they now use the proper length based on the
selected HMAC algorithm.