dh: workaround curve25519 'K' bug in OpenSSH
Workaround for curve25519 bug in OpenSSH:

  * https://red.libssh.org/issues/159

For OpenSSH clients 6.5p1 or 6.6p1, ensure that bignum strings
used for the shared-secret 'K' value in key exchange with
curve25519-sha256@libssh.org are indeed 32 bytes.

Before this change, OpenSSH clients starting at version 6.5p1 will
sometimes, but not always, error out in the early key exchange
phase along the lines of:

    debug2: kex_parse_kexinit: curve25519-sha256@libssh.org,ecdh-sha2-nistp256,diffie-hellman-group14-sha1,diffie-hellman-group1-sha1
    ...
    debug1: sending SSH2_MSG_KEX_ECDH_INIT
    debug1: expecting SSH2_MSG_KEX_ECDH_REPLY
    ...
    Warning: Permanently added '[localhost]:1234' (RSA) to the list of known hosts.
    hash mismatch
    debug1: ssh_rsa_verify: signature incorrect
    key_verify failed for server_host_key

After some digging it looks like there are some cases where the
ssh_string computed by libssh for its 'K' value does not have a
length of 32 bytes.  It is in these cases that OpenSSH clients
will fail signature verification during initial key exchange.
This is actually a bug in OpenSSH.

To workaround, for 6.5p1 and 6.6p1 clients, use a new helper for generating
'K' strings which inputs both a bignum and minimal size.  For the bad client
case, the 32 byte value is provided for the size; other key exchange
algorithms continue to use plain 'make_bignum_string' as before.

BUG: https://red.libssh.org/issues/159

Signed-off-by: Jon Simons <jon@jonsimons.org>